python/backend/pyproject.toml
[tool.poetry]
name = "vk-stats"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10.2"
aiohttp = "^3.8.4"
pydantic = "^2.0"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
asyncpg = "^0.29.0"
httpx = "^0.27.0"
pendulum = "^3.0.0"
structlog = "^24.1.0"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
python-dotenv = "^1.0.1"
prophet = "^1.1.5"
pandas = "^2.2.2"
numpy = "^1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
ruff-lsp = "^0.0.53"
pyright = "^1.1.363"
pynvim = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
extend-select = ["I"]

[tool.pyright]
include = ["src"]
executionEnvironments = [{ root = "src" }]
